:recycle: モデルをパッケージ化して、各モデルを別々のモジュールにした
