+++ src/models/__init__.py
@@ -1,0 +1,4 @@
+from .model_dt import ModelDecisionTreeClassifier
+from .model_lr import ModelLogisticRegression
+from .model_rf import ModelRandomForestClassifier
+from .model_xgb import ModelXgb
