+++ src/train.py
@@ -1,9 +1,6 @@
 # -*- coding: utf-8 -*-
 import argparse
-import os
-import joblib
 import pandas as pd
-from sklearn import metrics
 
 import config
 import model_dispatcher
